move prototypes into posix folder
